requirements.txt
discord==1.7.3
Pillow==9.0.0
requests==2.27.1